refactor: rename file to agent-worker-builder
